Former-commit-id: 0bf768a54b7787ad18f116782b170ebd8c35769b

+++ Sources/Internal/UI/UIStaticText.h
@@ -128,7 +128,7 @@ public:
     virtual Animation* TextColorAnimation(const Color& finalColor, float32 time, Interpolation::FuncType interpolationFunc = Interpolation::LINEAR, int32 track = 0);
     virtual Animation* ShadowColorAnimation(const Color& finalColor, float32 time, Interpolation::FuncType interpolationFunc = Interpolation::LINEAR, int32 track = 1);
 
-    const Vector<int32>& GetStringSizes() const;
+    const Vector<float32>& GetStringSizes() const;
 
     inline bool IsForceBiDiSupportEnabled() const
     {
